maks/requirements.txt
asgiref==3.3.1
certifi==2020.12.5
cffi
chardet==4.0.0
cryptography==3.4.6
defusedxml==0.6.0
Django==3.1.7
django-allauth==0.44.0
django-ckeditor==6.0.0
django-js-asset==1.2.2
django-wkhtmltopdf==3.3.0
idna==2.10
oauthlib==3.1.0
Pillow==8.1.1
pycparser==2.20
PyJWT==2.0.1
python3-openid==3.2.0
pytz==2021.1
requests==2.25.1
requests-oauthlib==1.3.0
six==1.15.0
sqlparse==0.4.1
urllib3==1.26.3
